LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/getstartpositionofchar-expected.txt

ABC
ABC
ABC
ABC
ABC

FAIL vertical-rl assert_equals: expected 113 but got 103
FAIL vertical-lr assert_equals: expected 113 but got 103
FAIL sideways-rl assert_equals: expected 113 but got 103
FAIL sideways-lr assert_equals: expected 97 but got 107
PASS tspan
